requirements.txt
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
future==0.18.2
iso8601==0.1.14
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow>=8.3.2
Pygments==2.9.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.2
six==1.15.0
